+++ src/main/java/uk/gov/justice/laa/amend/claim/interceptors/MaintenanceInterceptor.java
@@ -2,9 +2,11 @@ package uk.gov.justice.laa.amend.claim.interceptors;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
+import uk.gov.justice.laa.amend.claim.service.MaintenanceService;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -14,21 +16,20 @@ import java.util.Arrays;
 
 @Component
 @Slf4j
+@RequiredArgsConstructor
 public class MaintenanceInterceptor implements HandlerInterceptor {
+
+    private final MaintenanceService maintenanceService;
+
     @Override
     public boolean preHandle(
             HttpServletRequest request,
             HttpServletResponse response,
             Object handler) throws IOException {
-
-        if (hasBypassCookie(request)) {
-            return true;
-        }
-
         String path = request.getRequestURI();
         log.info("MaintenanceInterceptor path: {}", path);
 
-        if (!maintenanceEnabled()) {
+        if (!maintenanceService.maintenanceApplies(request)) {
             log.info("Maintenance off, allow: {}", path);
             return true;
         }
@@ -38,20 +39,5 @@ public class MaintenanceInterceptor implements HandlerInterceptor {
         return false;
     }
 
-    private boolean hasBypassCookie(HttpServletRequest request) {
-        return Arrays.stream(request.getCookies()).anyMatch(cookie -> cookie.getName().equals("maintenance_bypass"));
-    }
 
-    private boolean maintenanceEnabled() {
-        Path enabled = Paths.get("/config/maintenance/enabled");
-        try {
-            if (!Files.exists(enabled)) {
-                return false;
-            }
-            return Boolean.parseBoolean(Files.readString(enabled).trim());
-        } catch (IOException e) {
-            log.error("Failed to read maintenance flag", e);
-            return false;
-        }
-    }
 }
